itR-17: refactor: add castom hook useCityWeather
